Quick review with very minor changes
yahoo finance API no longer seems to be available, so fetching data from there is not possible anymore. Other than that, I quickly reviewed the programs. There are also issues with the way training was set up that has to do with time series and common features. I will address those in  a new revision, which is more like an academic paper with a neural network focus.